`DynamicSourceViewModel` should extend `BaseViewModel` instead of `ReactiveViewModel`
