+++ packages/stacked/test/futureviewmodel_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:stacked/stacked.dart';
+import 'package:stacked/src/state_management/multiple_data_models.dart';
+import 'package:stacked/src/state_management/single_data_models.dart';
 
 const _SingleFutureExceptionFailMessage = 'Future to Run failed';
