+++ packages/stacked/test/reactive_viewmodel_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:stacked/src/state_management/base_view_models.dart';
 import 'package:stacked/src/state_management/reactive_service_mixin.dart';
+import 'package:stacked/src/state_management/single_data_models.dart';
 
 class TestReactiveService with ReactiveServiceMixin {
   int _counter = 0;
