+++ packages/stacked/test/streamviewmodel_test.dart
@@ -1,4 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:stacked/src/state_management/multiple_data_models.dart';
+import 'package:stacked/src/state_management/single_data_models.dart';
 import 'package:stacked/stacked.dart';
 
 Stream<int> numberStream(int dataBack,
